+++ repyability/__init__.py
@@ -25,6 +25,7 @@ from repyability.rbd.repeated_standby_node import RepeatedStandbyNode
 from repyability.rbd.results import (
     AvailabilityResult,
     ConfidenceInterval,
+    CostResult,
     Criticalities,
     FailureCriticalityIndex,
     RestorationCriticalityIndex,
@@ -61,6 +62,7 @@ __all__ = [
     # Result types
     "AvailabilityResult",
     "ConfidenceInterval",
+    "CostResult",
     "Criticalities",
     "UpDownImportance",
     "FailureCriticalityIndex",